Hook search name refactor
