+++ src/Hooks/Search/Search.jsx
@@ -3,7 +3,7 @@ import React, { createContext, useContext, useState } from 'react';
 const SearchStateContext = createContext();
 const SearchDispatchContext = createContext();
 
-function SearchProvider({ children }) {
+function Search({ children }) {
   const [search, setSearch] = useState('');
   return (
     <SearchStateContext.Provider value={search}>
@@ -16,15 +16,15 @@ function SearchProvider({ children }) {
 function useSearchState() {
   const context = useContext(SearchStateContext);
   if (context === undefined) {
-    throw new Error('useSearchState must be used within a SearchProvider');
+    throw new Error('useSearchState must be used within a Search');
   }
   return context;
 }
 function useSearchDispatch() {
   const context = useContext(SearchDispatchContext);
   if (context === undefined) {
-    throw new Error('useSearchDispatch must be used within a SearchProvider');
+    throw new Error('useSearchDispatch must be used within a Search');
   }
   return context;
 }
-export { SearchProvider, useSearchState, useSearchDispatch };
+export { Search, useSearchState, useSearchDispatch };
