+++ src/components/App/App.component.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { BrowserRouter } from 'react-router-dom';
-import { SearchProvider } from '../../Hooks/SearchProvider/SearchProvider';
+import { Search } from '../../Hooks/Search/Search';
 import { Theme } from '../../Hooks/Theme/Theme';
 import Layout from '../Layout';
 import Routes from '../Routes/Routes.component';
@@ -11,11 +11,11 @@ function App() {
     <Theme>
       <GlobalStyles />
       <BrowserRouter>
-        <SearchProvider>
+        <Search>
           <Layout>
             <Routes />
           </Layout>
-        </SearchProvider>
+        </Search>
       </BrowserRouter>
     </Theme>
   );
